+++ io-sim/io-sim.cabal
@@ -98,11 +98,11 @@ test-suite test
   type:                exitcode-stdio-1.0
   hs-source-dirs:      test
   main-is:             Main.hs
-  other-modules:       Test.Control.Monad.STM
+  other-modules:       Test.Control.Concurrent.Class.MonadMVar
+                       Test.Control.Monad.STM
                        Test.Control.Monad.Utils
                        Test.Control.Monad.IOSim
                        Test.Control.Monad.IOSimPOR
-                       Test.Control.Monad.Class.MonadMVar
   default-language:    Haskell2010
   build-depends:       base,
                        array,
