+++ io-sim/test/Main.hs
@@ -2,7 +2,7 @@ module Main (main) where
 
 import           Test.Tasty
 
-import qualified Test.Control.Monad.Class.MonadMVar (tests)
+import qualified Test.Control.Concurrent.Class.MonadMVar (tests)
 import qualified Test.Control.Monad.IOSim (tests)
 import qualified Test.Control.Monad.IOSimPOR (tests)
 
@@ -12,7 +12,7 @@ main = defaultMain tests
 tests :: TestTree
 tests =
   testGroup "IO Sim"
-  [ Test.Control.Monad.Class.MonadMVar.tests
+  [ Test.Control.Concurrent.Class.MonadMVar.tests
   , Test.Control.Monad.IOSim.tests
   , Test.Control.Monad.IOSimPOR.tests
   ]
